requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
Werkzeug==2.0.1
requests==2.26.0
pandas==1.3.3
python-jose==3.3.0
bcrypt==3.2.0
